# src/core/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ApiBoiler"
  },
  "basePath": "/",
  "definitions": {
    "FishLimited": {
      "type": "object",
      "properties": {
        "ScientificName": {
          "type": ["string", "undefined"]
        },
        "A3Code": {
          "type": ["string", "undefined"]
        },
        "EnglishName": {
          "type": ["string", "undefined"]
        },
        "Nickname": {
          "type": ["string", "undefined"]
        },
        "Isscaap": {
          "type": ["number", "undefined"]
        },
        "Taxocode": {
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "Username": {
          "type": "string"
        },
        "Email": {
          "type": "string"
        },
        "PasswordHash": {
          "type": "string"
        },
        "PhoneNumber": {
          "type": ["string", "undefined"]
        },
        "Description": {
          "type": ["string", "undefined"]
        },
        "Name": {
          "type": ["string", "undefined"]
        }
      }
    },
    "UsernamePassword": {
      "type": "object",
      "properties": {
        "Username": {
          "type": "string"
        },
        "Password": {
          "type": "string"
        }
      }
    }
  },
  "paths": {
    "/user/login": {
      "post": {
        "summary": "Login a user",
        "tags": ["User"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "schema": {
              "$ref": "#/definitions/UsernamePassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful login",
            "schema": {
              "Token": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["User"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful registration",
            "schema": {
              "Token": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/user/update": {
      "post": {
        "summary": "Update a user's information",
        "tags": ["User"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User information updated",
            "schema": {
              "Token": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/user/delete": {
      "get": {
        "summary": "Delete a user's account",
        "tags": ["User"],
        "produces": ["text/plain"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "User account deleted"
          },
          "404": {
            "description": "No user to delete"
          }
        }
      }
    },
    "/worldfish/all": {
      "get": {
        "summary": "Get all fish from the database",
        "tags": ["World fish"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "FishInfo",
            "schema": {
              "$ref": "#/definitions/FishLimited"
            }
          }
        }
      }
    }
  }
}